SubjectクラスがTimeLimitクラスを継承するよう修正
