+++ src/main/java/jp/co/esm/novicetimer/api/TimerRestController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import jp.co.esm.novicetimer.domain.Subject;
+import jp.co.esm.novicetimer.domain.TimeLimit;
 import jp.co.esm.novicetimer.service.TimerService;
 
 @RestController
@@ -21,8 +21,8 @@ public class TimerRestController {
 
     @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
-    public String postTimers(@RequestBody Subject subject) {
-        return timerService.startTimer(subject);
+    public String postTimers(@RequestBody TimeLimit timeLimit) {
+        return timerService.startTimer(timeLimit);
     }
 
     @DeleteMapping
